// packages/shared-types/src/generated/openapi-document.ts
export const openApiDocument = {
// ...
  "paths": {
    "/v1/checkins": {
      "post": {
// ...
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInSubmission"
              }
            }
          }
        },
// ...
      }
    },
// ...
  },
// ...
} as const;